renaming to main and loadFromICE
